Float32 audio pipeline: end-to-end from radio to speaker
Replace int16 PCM with float32 throughout the entire RX/TX audio
pipeline. The radio sends float32 samples that can exceed 1.0 at
high AF levels — the old int16 path hard-clipped these, causing
distortion above AF slider ~70.

Changes:
- PanadapterStream: emit native float32 (no clamping/conversion)
- AudioEngine: QAudioFormat::Float sink, float32 DSP pipeline
- Resampler: float32 I/O (r8brain still uses double internally)
- RNNoiseFilter: remove int16↔float boundary conversions
- SpecbleachFilter: remove int16↔float boundary conversions
- DeepFilterFilter: remove int16↔float boundary conversions
- SpectralNR: float* I/O (was int16*)
- BNR processBnr: remove int16↔float conversions
- TciServer: float32 RX/TX audio, convert to int16 only for
  clients requesting format=0
- RADEEngine: float32 at resampler boundaries, int16 only at
  LPCNet/Opus API calls
- TX path: convert mic int16→float32 once at capture boundary
- computeRMS: simplified (no /32768 normalization needed)

+++ src/core/AudioEngine.h
@@ -272,8 +272,8 @@ private:
     std::atomic<bool> m_dfnrEnabled{false};
 
     // Pre-allocated NR2 work buffers (avoid per-call heap allocation)
-    std::vector<int16_t> m_nr2Mono;
-    std::vector<int16_t> m_nr2Processed;
+    std::vector<float> m_nr2Mono;
+    std::vector<float> m_nr2Processed;
     QByteArray m_nr2Output;
 
     // RX audio buffer handling
